requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
scipy
kneed
adjustText
pyyaml
statsmodels